feat: adding WEPRFeatureExtractor, sklearn transformer
